+++ server/services/jobService/createJob.js
@@ -1,0 +1,4 @@
+export default function createJob(jobQueueName, jobPayload, jobOptions) {
+  const queueService = require('src/server/services/queueService')
+  return queueService.getQueue(jobQueueName).add(jobPayload, jobOptions)
+}
